pjt01/back/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
